Cover vertical fanout handoff and diagnose CI failures
